% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\docType{data}
\name{Alaria}
\alias{Alaria}
\title{Test dataset for Alaria esculenta from uk (NBN Atlas 2021).}
\format{
dataframe object with columns Name, Latitude, Longitude, Year & Abundance with data for Alaria esculenta from UK
}
\usage{
data(Alaria)
}
\description{
Data frame containing Latitude, Longitude, Year & Abundance for Alaria esculenta from UK (NBN Atlas 2021)
}
\references{
Joint Nature Conservation Committee (2023). Marine Nature Conservation Review (MNCR) and associated benthic marine data held and managed by JNCC. Occurrence dataset on the NBN Atlas.
}
\keyword{datasets}
