% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{EvaluateA2b}
\alias{EvaluateA2b}
\title{Evaluate criterion A2b}
\usage{
EvaluateA2b(
  df,
  StartYear = NA,
  EndYear = NA,
  AssessmentYears = 10,
  MinGrids = 5
)
}
\arguments{
\item{df}{dataframe containing columns Latitude, Longitude, Grid (grid reference), Year (of observation) and Abundance (SACFOR category)}

\item{StartYear}{observation year to start the assessment}

\item{EndYear}{observation year to end the assessment (defaults to most recent year in data)}

\item{AssessmentYears}{number of years over which to perform the assessment (default 10)}

\item{MinGrids}{minimum number of repeat observations required to perform an assessment (default 5)}
}
\value{
A list with the following elements
Data: A data frame with columns GridRef, Year & Abundance (SACFOR category)
Trend: Observed trend in abundance (category change per year)
Trend.Percentiles: Percentils of observed trends
Percent: Percentage change in abundance over the assessment period based on the trend
NGrids: Number of grid cells with repeat surveys used for assessment
NObs: Number of observations used in the assessment (includes multiple observations at each grid cell)
A2b: threat category based on A2b criterion thresholds
}
\description{
Evaluate criteria A2b based on SACFOR abundance indices
}
\details{
EvaluateA2b() Measure change in abudance based on SACFORN index
compare change over the assessment period with A2b criterion thresholds
}
\examples{
data(Alaria)
a2b <- EvaluateA2b(Alaria)
}
