% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{EvaluateA2c}
\alias{EvaluateA2c}
\title{Evaluate criterion A2c}
\usage{
EvaluateA2c(df, StartYear = NA, EndYear = NA, AssessmentYears = 10)
}
\arguments{
\item{df}{dataframe containing columns Latitude, Longitude, Year (of observation)}

\item{StartYear}{observation year to start the assessment}

\item{EndYear}{observation year to end the assessment (defaults to most recent year in data)}

\item{AssessmentYears}{number of years over which to perform the assessment (default 10)}
}
\value{
A list containing the following elements
Data: a data frame with columns Year, AOO, EOO
AOO: a list with the following elements
A2c: a2c evaluation based on AOO
Area.AllTime: AOO based on all observations
Area.AllAssessment: AOO over the assessment period
Area.AssessmentStart: AOO at the first assessment year based on regression model
Area.AssessmentEnd: AOO at the final assessment year based on regression model
Slope: Slope of the regression Year~AOO (category decline per year)
p: significance value of the regression Year~AOO
Change.Percent: Percentage area change from the start to the end of the assessment period
EOO: a list with the following elements
A2c: a2c evaluation based on EOO
Area.AllTime: EOO based on all observations
Area.AllAssessment: EOO over the assessment period
Area.AssessmentStart: EOO at the first assessment year based on regression model
Area.AssessmentEnd: EOO at the final assessment year based on regression model
Slope: Slope of the regression Year~EOO (category decline per year)
p: significance value of the regression Year~EOO
Change.Percent: Percentage area change from the start to the end of the assessment period
}
\description{
Evaluate criterion A2c based on AOO & EOO trends derived from point observations
}
\details{
EvaluateA2c() employs a simple regression analysis of point distribution data
calculating area of occupancy and extent of occurrence values
to perform an assessment of distribution trend following Red List criterion A2c
}
\examples{
data(Alaria)
a2c <- EvaluateA2c(Alaria)
}
