% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{EvaluateAllCriteria}
\alias{EvaluateAllCriteria}
\title{Evaluate all Red List Criteria}
\usage{
EvaluateAllCriteria(
  df,
  StartYear = NA,
  EndYear = NA,
  AssessmentYears = 10,
  A2b.MinGrids = 5,
  D2.MinObs = 1,
  B.MinObs = 3
)
}
\arguments{
\item{df}{dataframe containing columns Latitude, Longitude, Year (of observation)}

\item{StartYear}{observation year to start the assessment}

\item{EndYear}{observation year to end the assessment (defaults to most recent year in data)}

\item{AssessmentYears}{number of years over which to perform the assessment (default 10)}

\item{A2b.MinGrids}{minimum number of repeat observations required to perform criterion A2b evaluation (default 5)}

\item{D2.MinObs}{minimum number of observations requires to perform criterion D2 evaluation (default 1)}

\item{B.MinObs}{minimum number of observations requires to perform criterion B evaluation (default 3)}
}
\value{
A list with the following elements
A2b: A list returned from function EvaluateA2b()
A2c: A list returned from function EvaluateA2c()
B: A list returned from function EvaluateB()
D2: A list returned from function EvaluateD2()
Final.Evaluation: The final evaluation based on all criteria
}
\description{
Evaluate all Red List criter from point observation data
}
\details{
EvaluateAllCriteria() Evaluate all Red List criteria by calling all criterion evaluations
}
\examples{
data(Alaria)
all <- EvaluateAllCriteria(Alaria)
}
