% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{EvaluateB}
\alias{EvaluateB}
\title{Evaluate criteria B1 & B2}
\usage{
EvaluateB(df, StartYear = NA, EndYear = NA, AssessmentYears = 10, MinObs = 3)
}
\arguments{
\item{df}{dataframe containing columns Latitude, Longitude, Year (of observation)}

\item{StartYear}{observation year to start the assessment}

\item{EndYear}{observation year to end the assessment (defaults to most recent year in data)}

\item{AssessmentYears}{number of years over which to perform the assessment (default 10)}

\item{MinObs}{minimum number of observations requires to perform an assessment (default 3)}
}
\value{
A list containing the following elements
AOO: AOO area based on the assessment period
AOOAllTime: AOO area over all time
AOO.Decline: Boolean indicating whether there is a long term AOO decline
B2.AOO: B2 threat category derived solely from AOO calculation
B2: final B2 threat category
EOO: EOO area based on the assessment period
EOOAllTime: EOO area over all time
B1.EOO: B1 threat category derived solely from EOO calculation
B1: final B1 threat category
NTetrads: Number of tetrads (2km x 2km blocks) with observations
NTetradBlocks: Number of contiguous areas based on tetrad presence
B.Locations: Number of locations based on the Tetrad proxy
StartYear: start year of evaluation
EndYear: end year of evaluation
}
\description{
Evaluate criteria B1 & B2 based on AOO & EOO derived from point observations
}
\details{
EvaluateB() calculate AOO & EOO from point distribution data
compare to B1 & B2 criteria thresholds and assign risk category
assess long term decline by comparing
}
\examples{
data(Alaria)
b <- EvaluateB(Alaria)
}
