% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{EvaluateD2}
\alias{EvaluateD2}
\title{Evaluate criterion D2}
\usage{
EvaluateD2(df, StartYear = NA, EndYear = NA, MinObs = 1)
}
\arguments{
\item{StartYear}{observation year to start the assessment}

\item{EndYear}{observation year to end the assessment (defaults to most recent year in data)}

\item{MinObs}{minimum number of observations requires to perform an assessment (default 1)}

\item{dataframe}{containing columns latitude, longitude, year (of observation)}
}
\value{
A list with the following elements
AOO: AOO area
D2: D2 threat category based solely on AOO area (no consideration of ongoing threats)
}
\description{
Evaluate criteria D2 based on AOO
}
\details{
EvaluateD2() calculate AOO from point distribution data and compares to area threshold D2
Warning! This makes no evaluation of ongoing threats. To formally apply the D2 criterion you must document a threat
}
\examples{
b <- EvaluateD2(Alaria)
}
