% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{PlotA2b}
\alias{PlotA2b}
\title{Plot data used for criterion A2b evaluation}
\usage{
PlotA2b(A2b)
}
\arguments{
\item{A2b}{a list output from EvaluateA2b()}
}
\value{
A ggplot object
}
\description{
Plot data used for criterion A2b evaluation
}
\details{
Creates a scatter plot of Abundance~Year used for A2b evaluation
}
\examples{
A2b <- EvaluateA2b(Alaria)
PlotA2b(A2b)
}
