% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{PlotA2c}
\alias{PlotA2c}
\title{Plot data used for criterion A2c evaluation}
\usage{
PlotA2c(A2c, type = "AOO")
}
\arguments{
\item{A2c}{a list output from EvaluateA2c()}

\item{type}{a character string either "AOO" or "EOO" indicationg which metric to plot}
}
\value{
A ggplot object
}
\description{
Plot data used for criterion A2c evaluation
}
\details{
Creates a scatter plot of Area~Year used for A2c evaluation
}
\examples{
data(Alaria)
A2c <- EvaluateA2c(Alaria)
PlotA2c(A2c)
}
