% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedListEvaluation.R
\name{PlotArea}
\alias{PlotArea}
\title{Plot map based on point observations}
\usage{
PlotArea(df, B, D2, CRS = NA, GridRes = 2000)
}
\arguments{
\item{df}{a data frame used in evaluations}

\item{B}{a list returned from EvaluateB()}

\item{D2}{a list returned from EvaluateD2()}

\item{CRS}{a coordinate reference system definition recognisable by sf::st_crs - will guess a local UTM zone if left blank}

\item{GridRes}{a value indicating resolution}
}
\value{
A ggplot object
}
\description{
Plot map based on point observations used in evaluation
}
\details{
Plot a simple map based on unique tetrads of point observations used in evaluation
display results of B1&2 and D2 evaluations too
}
\examples{
data(Alaria)
B <- EvaluateB(Alaria)
D2 <- EvaluateD2(Alaria)
PlotArea(Alaria, B, D2)
}
